minor fixes for narrowing errors

+++ src/libmerc/cbor.hpp
@@ -78,7 +78,7 @@ namespace cbor {
         // construct an initial_byte for writing
         //
         initial_byte(uint8_t type, uint8_t info) :
-            value__{type << 5 | info}
+            value__{(uint8_t)(type << 5 | info)}
         {
             // printf("major_type: %u\n", major_type());
             // printf("additional_info: %u\n", additional_info());
@@ -222,16 +222,16 @@ namespace cbor {
             ib.write(buf);
             switch (ib.additional_info()) {
             case 24:
-                encoded<uint8_t>{value__}.write(buf, true);
+                encoded<uint8_t>{(uint8_t)value__}.write(buf, true);
                 break;
             case 25:
-                encoded<uint16_t>{value__}.write(buf, true);
+                encoded<uint16_t>{(uint16_t)value__}.write(buf, true);
                 break;
             case 26:
-                encoded<uint32_t>{value__}.write(buf, true);
+                encoded<uint32_t>{(uint32_t)value__}.write(buf, true);
                 break;
             case 27:
-                encoded<uint64_t>{value__}.write(buf, true);
+                encoded<uint64_t>{(uint64_t)value__}.write(buf, true);
                 break;
             default:
                 ;
@@ -313,7 +313,7 @@ namespace cbor {
         ///
         static byte_string decode(datum &d) {
             uint64 len{d, byte_string_type};
-            datum val{d, len.value()};
+            datum val{d, (ssize_t)len.value()};
             return byte_string{len, val};
         }
 
@@ -321,7 +321,7 @@ namespace cbor {
         /// bytes in the \ref datum \param d
         ///
         static byte_string construct(const datum &d) {
-            uint64 len{d.length(), byte_string_type};
+            uint64 len{(uint64_t)d.length(), byte_string_type};
             datum val{d};
             return byte_string{len, val};
         }
@@ -365,7 +365,7 @@ namespace cbor {
     public:
 
         byte_string_from_hex(const hex_digits &hex_string) :
-            length{hex_string.length() / 2, byte_string_type},
+            length{((uint64_t)hex_string.length()) / 2, byte_string_type},
             hex_value{hex_string}
         { }
 
@@ -422,7 +422,7 @@ namespace cbor {
         ///
         static text_string decode(datum &d) {
             uint64 len{d, text_string_type};
-            datum val{d, len.value()};
+            datum val{d, (ssize_t)len.value()};
             return text_string{len, val};
         }
 
@@ -430,7 +430,7 @@ namespace cbor {
         /// bytes in the \ref datum \param d
         ///
         static text_string construct(const datum &d) {
-            uint64 len{d.length(), text_string_type};
+            uint64 len{(uint64_t)d.length(), text_string_type};
             datum val{d};
             return text_string{len, val};
         }
@@ -450,7 +450,7 @@ namespace cbor {
         // construct a text_string for writing
         //
         text_string(const char *null_terminated_string) :
-            length{datum_from_str(null_terminated_string).length(), text_string_type},
+            length{(uint64_t)(datum_from_str(null_terminated_string).length()), text_string_type},
             value__{datum_from_str(null_terminated_string)}
         { }
